refactor: extract OAuthError to types/error.ts, move validateRedirectUri to utils/redirect.ts
